+++ tests/BaseTest.php
@@ -2,7 +2,6 @@
 namespace MallardDuck\Whois\Test;
 
 use PHPUnit\Framework\TestCase;
-use League\Uri\Components\Host;
 
 /**
 *  Corresponding Class to test the whois Client class
@@ -14,15 +13,14 @@ use League\Uri\Components\Host;
 */
 abstract class BaseTest extends TestCase
 {
+
+    /**
+     * [getMethod description]
+     * @return \League\Uri\Components\Exception         [description]
+     */
     public function getUriException()
     {
-        $host = new Host();
-        $isNewLib = (method_exists($host, 'getRegistrableDomain')) ? true : false;
-        if ($isNewLib) {
-            return \League\Uri\Components\Exception::class;
-        }
-
-        return \Exception::class;
+        return \League\Uri\Components\Exception::class;
     }
 
     /**